Add tooling to dump customfield menu data
JIRA dropdown field value menus are yet another aggravating sharp
edge we have to work around.

If you have a predefined list of menu items, you can't just shovel
in a string that corresponds to one of those defined menu items.
JIRA isn't smart enough to compare that string to its list of
allowed values and use it when it's a valid option.

Instead, you have to figure out what id that value corresponds to,
and set _that_. Along with stuffing in the original value, of course,
Atlassian wouldn't want this to be convenient in any way.

So to work around that, we have to be able to create a lookup table
from a ticket. So we're adding `dump_all_customfield_allowed_values()`
to the JIRA class and a supporting cli command to dump it when we're
debugging.

Suck abounds.

Signed-off-by: Joe Block <jblock@zscaler.com>