Fix bad argument handling in jc-create-ticket
Fix bad handling of unset cli arguments. When unset, arpgarse
defaults them to None, passing them to JIRA makes it puke up errors.

Found this when the PAT from the settings file would be overwritten
by the cli pat-token argument, even if that was None, which is the
default value returned by argparse when an argument is unset on the
command line.

Signed-off-by: Joe Block <jblock@zscaler.com>